+++ packages/chain-adapters/src/bitcoin/BitcoinChainAdapter.ts
@@ -1,4 +1,4 @@
-import { AssetNamespace, AssetReference, CAIP2, caip2, caip19 } from '@shapeshiftoss/caip'
+import { AssetNamespace, AssetReference, CAIP2, fromCAIP2, toCAIP19 } from '@shapeshiftoss/caip'
 import {
   bip32ToAddressNList,
   BTCOutputAddressType,
@@ -56,12 +56,12 @@ export class ChainAdapter
     } else {
       this.chainId = this.supportedChainIds[0]
     }
-    const { chain, network } = caip2.fromCAIP2(this.chainId)
+    const { chain, network } = fromCAIP2(this.chainId)
     if (chain !== ChainTypes.Bitcoin) {
       throw new Error('chainId must be a bitcoin chain type')
     }
     this.coinName = args.coinName
-    this.assetId = caip19.toCAIP19({
+    this.assetId = toCAIP19({
       chain,
       network,
       assetNamespace: AssetNamespace.Slip44,
@@ -78,8 +78,8 @@ export class ChainAdapter
   }
 
   async getTxHistory(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    input: chainAdapters.TxHistoryInput
+    // @ts-ignore: keep type signature with unimplemented state
+    input: chainAdapters.TxHistoryInput // eslint-disable-line @typescript-eslint/no-unused-vars
   ): Promise<chainAdapters.TxHistoryResponse<ChainTypes.Bitcoin>> {
     throw new Error('Method not implemented.')
   }