+++ packages/asset-service/CHANGELOG.md
@@ -1,3 +1,19 @@
+# [@shapeshiftoss/asset-service-v2.6.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.5.0...@shapeshiftoss/asset-service-v2.6.0) (2022-05-06)
+
+
+### Features
+
+* **caip:** flatten exports ([#560](https://github.com/shapeshift/lib/issues/560)) ([e326522](https://github.com/shapeshift/lib/commit/e3265223dca3c2126b2822395353f6650c4b0342))
+
+# [@shapeshiftoss/asset-service-v2.5.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.4.6...@shapeshiftoss/asset-service-v2.5.0) (2022-05-06)
+
+
+### Features
+
+* regenerate asset data with assetId as sorting criteria ([#620](https://github.com/shapeshift/lib/issues/620)) ([716e2e1](https://github.com/shapeshift/lib/commit/716e2e13847f58eadd5ec90dbd0d2c52db6d62cf))
+
+# [@shapeshiftoss/asset-service-v2.4.6](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.4.5...@shapeshiftoss/asset-service-v2.4.6) (2022-05-05)
+
 # [@shapeshiftoss/asset-service-v2.4.5](https://github.com/shapeshift/lib/compare/@shapeshiftoss/asset-service-v2.4.4...@shapeshiftoss/asset-service-v2.4.5) (2022-05-03)